refactor(unntaksvurdering): del opp meldeskjema og stram kontrakten
