+++ src/components/OversiktSide/MeldUnntak/__tests__/MeldUnntakSection.test.tsx
@@ -89,7 +89,6 @@ describe("MeldUnntakSection", () => {
     expect(
       screen.getByText(/Du må rette dette før du kan sende/i),
     ).toBeInTheDocument();
-    // Fokus flyttes til ErrorSummary (WCAG: feil annonseres for skjermleser).
     await waitFor(() =>
       expect(document.activeElement?.textContent).toMatch(
         /Du må rette dette før du kan sende/i,
@@ -98,7 +97,7 @@ describe("MeldUnntakSection", () => {
     expect(mockMeldUnntak).not.toHaveBeenCalled();
   });
 
-  test("kaller server action og viser kvittering med skissens tekst", async () => {
+  test("kaller server action og viser kvittering", async () => {
     mockMeldUnntak.mockResolvedValue({ error: null });
     renderMeldUnntak();
 
@@ -110,8 +109,6 @@ describe("MeldUnntakSection", () => {
 
     expect(mockMeldUnntak).toHaveBeenCalledWith("test-leder-id");
 
-    // Hele flaten er gated bak big bang-toggelen, så teksten kan love
-    // sykmeldt-visningen (#888) — alt lanseres samlet.
     expect(
       await screen.findByText(/Meldingen er sendt til Nav og den ansatte/i),
     ).toBeInTheDocument();
@@ -120,7 +117,6 @@ describe("MeldUnntakSection", () => {
         /Nav har registrert at det ikke er aktuelt med en oppfølgingsplan for Kreativ Hatt nå/i,
       ),
     ).toBeInTheDocument();
-    // Skjemaet er erstattet av kvitteringen.
     expect(
       screen.queryByRole("button", { name: /Send til Nav og den ansatte/i }),
     ).not.toBeInTheDocument();
@@ -142,7 +138,6 @@ describe("MeldUnntakSection", () => {
     expect(
       screen.queryByText(/Meldingen er sendt til Nav og den ansatte/i),
     ).not.toBeInTheDocument();
-    // Arbeidsgiver kan melde på nytt — valget består etter meldt unntak.
     expect(
       screen.getByRole("heading", {
         name: /Det finnes noen unntak fra å lage oppfølgingsplan/i,
