+++ src/schema/oversiktResponseSchemas.ts
@@ -25,16 +25,8 @@ export const OppfolgingsplanerOversiktResponseSchemaForAG = z.object({
     utkast: utkastMetadataSchema.nullable(),
     aktivPlan: oppfolgingsplanMetadataSchema.nullable(),
     tidligerePlaner: z.array(oppfolgingsplanMetadataSchema),
-    /**
-     * Meldte unntaksvurderinger, nyeste først. Egen liste i kontrakten —
-     * skal aldri blandes inn i tidligerePlaner (gatingen av «Lag ny plan»
-     * skal ikke påvirkes av meldte unntak).
-     *
-     * default/optional til backend-PR navikt/syfo-oppfolgingsplan-backend#400
-     * er ute i prod; kan da strammes til påkrevd.
-     */
-    unntaksvurderinger: z.array(unntaksvurderingMetadataSchema).default([]),
-    gjeldendeStatus: gjeldendeStatusSchema.optional(),
+    unntaksvurderinger: z.array(unntaksvurderingMetadataSchema),
+    gjeldendeStatus: gjeldendeStatusSchema,
   }),
 });
