+++ src/server/actions/meldUnntaksvurdering.ts
@@ -8,11 +8,6 @@ import { simulateBackendDelay } from "../fetchData/mockData/simulateBackendDelay
 import type { FetchUpdateResult } from "../tokenXFetch/FetchResult";
 import { tokenXFetchUpdate } from "../tokenXFetch/tokenXFetchUpdate";
 
-/**
- * Melder at oppfølgingsplan ikke er aktuell nå (unntaksvurdering).
- * Ingen request-body — backend lagrer kun hvem og når.
- * Backend avviser med 409 hvis det finnes aktiv plan eller utkast.
- */
 export async function meldUnntaksvurderingServerAction(
   narmesteLederId: string,
 ): Promise<FetchUpdateResult> {
