+++ src/server/fetchData/sykmeldt/fetchOppfolgingsplanOversiktForSM.ts
@@ -22,7 +22,6 @@ import { simulateBackendDelay } from "../mockData/simulateBackendDelay";
 export function getMockDataForScenarioSM(scenario: DemoScenario) {
   switch (scenario) {
     case "tom":
-    // Sykmeldt-visningen av unntak bygges i #888 — til da ser sykmeldt tom liste.
     case "unntak-meldt":
       return mockOversiktDataTomForSM;
     case "aktiv-og-tidligere":
